+++ test/integration/faculty/app_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../helpers'
-
-describe file('/home/django/.ssh/known_hosts') do
-  # Basics tested in default suite
-  its(:content) { should match(/github\.com ssh-rsa/) }
-end