+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog for Django Platform Cookbook
 
+## 1.0.0
+
+* Parallelized test cases
+* Added support for multiple git hosts
+* Added hooks for recipes
+* Added flag for repo update
+
 ## 0.1.0
 
 * Initial release